requirements.txt
huggingface-hub==0.29.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
